refactor(react): complete media synchronization separation
